finalizing the project and deleting everything that not needed
