+++ Ex1/src/algorithem/Graph.java
@@ -1,7 +1,8 @@
+package algorithem;
 
+import java.util.Iterator;
 import java.util.Stack;
 
-import algorithem.DirectedEdge;
 import outclasses.Bag;
 import outclasses.In;
 import outclasses.StdOut;
@@ -66,7 +67,7 @@ public class Graph {
 
     private final int V;
     private int E;
-    private Bag<Integer>[] adj;
+    private Bag<DirectedEdge>[] adj;
     
     /**
      * Initializes an empty graph with <tt>V</tt> vertices and 0 edges.
@@ -80,9 +81,9 @@ public class Graph {
         if (V < 0) throw new IllegalArgumentException("Number of vertices must be nonnegative");
         this.V = V;
         this.E = 0;
-        adj = (Bag<Integer>[]) new Bag[V];
+        adj = (Bag<DirectedEdge>[]) new Bag[V];
         for (int v = 0; v < V; v++) {
-            adj[v] = new Bag<Integer>();
+            adj[v] = new Bag<DirectedEdge>();
         }
     }
 
@@ -103,7 +104,8 @@ public class Graph {
         for (int i = 0; i < E; i++) {
             int v = in.readInt();
             int w = in.readInt();
-            addEdge(v, w);
+            double weight = in.readDouble();
+            addEdge(new DirectedEdge(v, w, weight));
         }
     }
 
@@ -117,11 +119,11 @@ public class Graph {
         this.E = G.E();
         for (int v = 0; v < G.V(); v++) {
             // reverse so that adjacency list is in same order as original
-            Stack<Integer> reverse = new Stack<Integer>();
-            for (int w : G.adj[v]) {
+            Stack<DirectedEdge> reverse = new Stack<DirectedEdge>();
+            for (DirectedEdge w : G.adj[v]) {
                 reverse.push(w);
             }
-            for (int w : reverse) {
+            for (DirectedEdge w : reverse) {
                 adj[v].add(w);
             }
         }
@@ -158,12 +160,14 @@ public class Graph {
      * @param  w the other vertex in the edge
      * @throws IndexOutOfBoundsException unless both 0 <= v < V and 0 <= w < V
      */
-    public void addEdge(int v, int w) {
+    public void addEdge(DirectedEdge e) {
+    	int v = e.from();
+    	int w = e.to();
         validateVertex(v);
         validateVertex(w);
         E++;
-        adj[v].add(w);
-        adj[w].add(v);
+        adj[v].add(e);
+        
     }
 
 
@@ -174,10 +178,10 @@ public class Graph {
      * @return the vertices adjacent to vertex <tt>v</tt>, as an iterable
      * @throws IndexOutOfBoundsException unless 0 <= v < V
      */
-    public Iterable<Integer> adj(int v) {
+   /* public Iterable<Integer> adj(int v) {
         validateVertex(v);
         return adj[v];
-    }
+    }*/
 
     /**
      * Returns the degree of vertex <tt>v</tt>.
@@ -203,23 +207,62 @@ public class Graph {
         s.append(V + " vertices, " + E + " edges " + NEWLINE);
         for (int v = 0; v < V; v++) {
             s.append(v + ": ");
-            for (int w : adj[v]) {
+            for (DirectedEdge w : adj[v]) {
                 s.append(w + " ");
             }
             s.append(NEWLINE);
         }
         return s.toString();
     }
-
-
+    
     /**
-     * Unit tests the <tt>Graph</tt> data type.
-     */
-    public static void main(String[] args) {
-        In in = new In(args[0]);
-        Graph G = new Graph(in);
-        StdOut.println(G);
+	 * Returns the directed edges incident from vertex <tt>v</tt>.
+	 *
+	 * @param  v the vertex
+	 * @return the directed edges incident from vertex <tt>v</tt> as an Iterable
+	 * @throws IndexOutOfBoundsException unless 0 <= v < V
+	 */
+	public Iterable<DirectedEdge> adj(int v) {
+		validateVertex(v);
+		return adj[v];
+	}
+	
+    public Iterable<DirectedEdge> edges() {
+        Bag<DirectedEdge> list = new Bag<DirectedEdge>();
+        for (int v = 0; v < V; v++) {
+            for (DirectedEdge e : adj(v)) {
+                list.add(e);
+            }
+        }
+        return list;
+    } 
+    
+    public void setValidateVertex(int v, boolean value)
+    {
+      Iterable<DirectedEdge> ee = adj(v);
+      Iterator<DirectedEdge> i = ee.iterator();
+      while (i.hasNext())
+      {
+        DirectedEdge de = (DirectedEdge)i.next();
+        de.setValid(value);
+      }
     }
     
+    public void setValidateVertex(int[] vv, boolean value)
+    {
+      for (int i = 0; i < vv.length; i++) {
+        setValidateVertex(vv[i], value);
+      }
+    }
+    public void validAll()
+    {
+      Iterable<DirectedEdge> ee = edges();
+      Iterator<DirectedEdge> i = ee.iterator();
+      while (i.hasNext())
+      {
+        DirectedEdge de = (DirectedEdge)i.next();
+        de.setValid(true);
+      }
+    }
     
 }
