+++ Ex1/tests/Solution_test6EWG.txt_tinyEWG.txt.txt
